+++ src/modules/feature-modules/innovator/innovator.module.ts
@@ -26,21 +26,21 @@ import { PageInnovationExportRequestRejectComponent } from './pages/innovation/e
 import { PageInnovationHowToProceedComponent } from './pages/innovation/how-to-proceed/how-to-proceed.component';
 import { PageInnovationManageAccessLeaveInnovationComponent } from './pages/innovation/manage-access/manage-access-leave-innovation.component';
 import { PageInnovationManageAccessOverviewComponent } from './pages/innovation/manage-access/manage-access-overview.component';
-import { PageInnovationManageCollaboratorsInfoComponent } from './pages/innovation/manage/manage-collaborators-info.component';
-import { PageInnovationManageCollaboratorsOverviewComponent } from './pages/innovation/manage/manage-collaborators-overview.component';
-import { PageInnovationManageCollaboratorsWizardComponent } from './pages/innovation/manage/manage-collaborators-wizard.component';
-import { PageInnovationManageOverviewComponent } from './pages/innovation/manage/manage-overview.component';
-import { PageInnovationManageTransferComponent } from './pages/innovation/manage/manage-transfer.component';
+import { PageInnovationManageCollaboratorsInfoComponent } from './pages/innovation/settings/manage-collaborators-info.component';
+import { PageInnovationManageCollaboratorsOverviewComponent } from './pages/innovation/settings/manage-collaborators-overview.component';
+import { PageInnovationManageCollaboratorsWizardComponent } from './pages/innovation/settings/manage-collaborators-wizard.component';
+import { PageInnovationManageOverviewComponent } from './pages/innovation/settings/manage-overview.component';
+import { PageInnovationManageTransferComponent } from './pages/innovation/settings/manage-transfer.component';
 import { PageInnovationNeedsReassessmentSendComponent } from './pages/innovation/needs-reassessment/needs-reassessment-send.component';
 import { InnovationOverviewComponent } from './pages/innovation/overview/overview.component';
 import { InnovationDataSharingEditComponent } from './pages/innovation/record/data-sharing-edit.component';
 import { InnovationSectionEvidenceEditComponent } from './pages/innovation/record/evidence-edit.component';
 import { InnovationSectionEditComponent } from './pages/innovation/record/section-edit.component';
 import { InnovationSectionSubmittedComponent } from './pages/innovation/record/section-submitted.component';
-import { WizardInnovationManageArchiveComponent } from './pages/innovation/manage/wizard-manage-archive/manage-archive.component';
-import { WizardInnovationManageArchiveReasonStepComponent } from './pages/innovation/manage/wizard-manage-archive/steps/reason-step.component';
-import { WizardInnovationManageArchiveConfirmationStepComponent } from './pages/innovation/manage/wizard-manage-archive/steps/confirmation-step.component';
-import { WizardInnovationManageArchiveOverviewStepComponent } from './pages/innovation/manage/wizard-manage-archive/steps/overview-step.component';
+import { WizardInnovationManageArchiveComponent } from './pages/innovation/settings/wizard-manage-archive/manage-archive.component';
+import { WizardInnovationManageArchiveReasonStepComponent } from './pages/innovation/settings/wizard-manage-archive/steps/reason-step.component';
+import { WizardInnovationManageArchiveConfirmationStepComponent } from './pages/innovation/settings/wizard-manage-archive/steps/confirmation-step.component';
+import { WizardInnovationManageArchiveOverviewStepComponent } from './pages/innovation/settings/wizard-manage-archive/steps/overview-step.component';
 import { WizardInnovationHowToProceedArchiveComponent } from './pages/innovation/how-to-proceed/wizard-how-to-proceed-archive/how-to-proceed-archive.component';
 import { PageInnovationSupportSurveysComponent } from './pages/innovation/surveys/support-surveys.component';
 import { EndSupportSurveyJourneyComponent } from './pages/innovation/surveys/journeys/end-support.component';
