+++ src/modules/feature-modules/innovator/pages/innovation/settings/manage-collaborators-info.component.ts
