+++ src/modules/feature-modules/innovator/pages/innovation/settings/manage-collaborators-overview.component.html
